implemented auth

+++ src/routes/deleteListing.js
@@ -1,11 +1,15 @@
+import * as admin from 'firebase-admin'
 import { db } from "../database";
 
 export const deleteListingRoute = {
     method: "DELETE",
     path: "/api/listings/{id}",
     handler: async (req, h) => {
+        const token = req.headers.authtoken;
+        const user = await admin.auth().verifyIdToken(token)
+        const userId = user.user_id
         const { id } = req.params;
-        await db.query("delete from listings where id =?", [id]);
+        await db.query("delete from listings where id =? and user_id = ?", [id, userId]);
         return { message: "success" };
     },
 };
